Update README.md
This keyer can have a minimalist configuration with only a paddle required. Control parameters can be entered in Morse frm the paddle. But it can also be fitted with pushbottons for memories, a speed pot and a speaker for sidetone monitoring.
The keyer features four memories, A/B logic selection, straight key mode, pitch control, low battery warning and more.